Switched to alternative design of limiter
I've altered the way that the limiter works to make it a bit simpler for the consumer. Now what happens is that the consumer passes in a `Task` to be executed, and the limiter takes care of deciding whether or not to execute it.

I've also fixed a number of StyleCop warnings.