Initial implementation of metrics support
At the moment I've just created an abstraction, and instrumented the code. I've created some default implementations of the metrics interfaces that do nothing to support the use-case where no metrics are required.

In order to test the `concurrency_limiter_limited_total` metric, I had to create a situation where the current number of executing requests were enough to trigger the limit. To do that I ended up using a `ManualResetEvent` to create tasks that would block until signalled.